restructure search page WIP
- extract SearchPanel and FilterPanel
- restructure markup/CSS and apply Theme 0.1 styles
